+++ page-objects/pages/HMWRKHomeworkPage.js
@@ -19,15 +19,6 @@ const studentSubmissionTab = '#submission-tab-link';
 const submissionContainer = '.table .usersubmission';
 const remoteFilePathInput = 'input[type=file][class=dz-hidden-input]';
 const commentBtn = "a#comment-tab-link.tab-link";
-const dashboardTitleListSel = '.dashboard-title';
-
-const element = {
-	homeworkName: 'span[data-testid="homework-name"]',
-	courseName: 'span[data-testid="homework-course-name"]',
-	homeworkTimeout: 'span[data-testid="homework-due-date"]',
-	homeworkCompleted: 'span[data-testid="homework-submitted"]',
-	homeworkGraded: 'span[data-testid="homework-graded"]',
-};
 
 async function submitSolutionForTheHometask() {
 	const assignmentText = 'here is some text which I want to submit';
@@ -48,7 +39,6 @@ async function openStudentSubmissionTab() {
 	await elementHelpers.click(studentSubmissionTab);
 }
 
-// teacher helpers
 async function hasTheStudentSubmittedTheTask(studentname) {
 	await openTeacherSubmissionsTab();
 	const listOfSubmisionStudentNames = await getListOfSubmisions();
@@ -153,53 +143,6 @@ async function getCurrentTabUrl() {
 	return new URL(await driver.getUrl());
 }
 
-async function isPrivateTasksAndDraftsListVisible(listName) {
-	const dashboardTitlesList = await elementHelpers.getTextFromAllElements(dashboardTitleListSel)
-	const msg = 'Title with name [' + listName + '] is not visible on the list \n';
-	const resultMsg = ', List of titles: ' + dashboardTitlesList;
-	return expect(dashboardTitlesList, msg + resultMsg).to.include(listName);
-}
-
-async function isCourseNameOnPrivateHomeworkVisible(homeworkName, courseName) {
-	const homeworkIndex = await getIndexOfHomeworkFromList(homeworkName);
-	const homeworkCourseElementList = await elementHelpers.getListOfAllElements(element.courseName)
-	const courseNameText = await elementHelpers.getElementText(homeworkCourseElementList[homeworkIndex]);
-	const msg = 'Course with name [' + courseName + '] is not visible on the private homework \n';
-	const resultMsg = 'List of course titles: ' + homeworkCourseElementList;
-	return expect(courseNameText, msg + resultMsg).to.include(courseName);
-}
-
-async function getIndexOfHomeworkFromList(homeworkName) {
-	const homeworkTextElementList = await elementHelpers.getListOfAllElements(element.homeworkName)
-	const homeworkTextValueList = await elementHelpers.getTextListFromListOfElements(homeworkTextElementList);
-	const index = homeworkTextValueList.indexOf(homeworkName);
-	return index;
-}
-
-async function isPrivateHomeworkNameVisible(homeworkName) {
-	const homeworkNameList = await elementHelpers.getTextFromAllElements(element.homeworkName)
-	const msg = 'Homework with name [' + homeworkName + '] is not visible on the list \n';
-	const resultMsg = 'List of Homework titles: ' + homeworkNameList;
-	return expect(homeworkNameList, msg + resultMsg).to.include(homeworkName);
-}
-
-async function isElementOfHomeworkVisible(elementName, homeworkName, selector, expectedValue) {
-	const defaultString = `Element with name: ${elementName}`;
-	let elementOfHomework = true;
-	const homeworkIndex = await getIndexOfHomeworkFromList(homeworkName);
-	const elementsOfHomeworksList = await elementHelpers.getListOfAllElements(selector);
-	elementsOfHomeworksList.length === 0 ? elementOfHomework = false : await waitHelpers.waitUntilElementIsPresent(elementsOfHomeworksList[homeworkIndex]);
-
-	const msg = expectedValue
-		? `${defaultString} should be visible on the homework`
-		: `${defaultString} should not be visible on the homework`;
-
-	expectedValue
-		? expect(elementOfHomework, msg).to.be.true
-		: expect(elementOfHomework, msg).to.be.false;
-
-}
-
 module.exports = {
 	openSubmissionsTab: openTeacherSubmissionsTab,
 	submitSolutionForTheHometask,
@@ -212,10 +155,5 @@ module.exports = {
 	testFileUploadSuccess,
 	isFileVisible,
 	getCurrentTabUrl,
-	isPrivateTasksAndDraftsListVisible,
-	isCourseNameOnPrivateHomeworkVisible,
-	isPrivateHomeworkNameVisible,
-	isElementOfHomeworkVisible,
-	element,
 
 };
